fix(shields): harden unsafe Hermes path rollback (#8804)
Only clear provisional DOWN when unlock never completed, open preflight
files through a held config dir fd, and cover unlock-then-unsafe re-lock.
