+++ src/lib/shields/openclaw-transition.test.ts
@@ -8,6 +8,12 @@ import path from "node:path";
 
 import { afterEach, beforeEach, describe, expect, it, type MockInstance, vi } from "vitest";
 import {
+  createHermesUnsafeConfigHarness,
+  expectHermesShieldsUpRecord,
+  failHermesInferenceConvergence,
+  type HermesUnsafeConfigHarness,
+} from "../../../test/helpers/hermes-unsafe-config-shields-harness";
+import {
   createShieldsFlowHarness,
   type ShieldsFlowHarnessOptions,
 } from "../../../test/helpers/shields-flow-harness";
@@ -769,268 +775,71 @@ describe("OpenClaw shields flow rollback and recovery", () => {
 });
 
 describe("Hermes Shields down unsafe config path (#8804)", () => {
-  const HERMES_PYTHON = "/opt/hermes/.venv/bin/python";
-  const HERMES_GUARD = "/usr/local/lib/nemoclaw/hermes-runtime-config-guard.py";
-  const LOCK_TOKEN = "a".repeat(64);
-  const CURRENT_GUARD_HELP = [
-    "begin-shields-transition",
-    "run-state-dir-transition",
-    "apply-shields-transition",
-    "finish-shields-transition",
-    "prepare-shields-abort",
-    "abort-shields-transition",
-    "--rollback-shields-mode",
-    "--state-lock-plan-json",
-  ].join(" ");
-  const hermesTarget = {
-    agentName: "hermes",
-    configPath: "/sandbox/.hermes/config.yaml",
-    configDir: "/sandbox/.hermes",
-    format: "yaml",
-    configFile: "config.yaml",
-    sensitiveFiles: ["/sandbox/.hermes/.env", "/sandbox/.hermes/.config-hash"],
-    stateLockPlan: {
-      version: 1 as const,
-      readOnlyRoots: ["skills"],
-      confidentialRoots: ["pairing"],
-      readOnlyPrefixes: [],
-      confidentialPrefixes: [],
-      writableSubpaths: [],
-    },
-    stateLockPlanInImage: true,
-  };
-
-  let homeDir: string;
-  let shields: typeof import("./index.js");
-  let runSpy: MockInstance;
-  let dockerExecSpy: MockInstance;
-  let auditSpy: MockInstance;
-  let errorSpy: MockInstance;
-
-  function isGuardAction(cmd: string[], action: string): boolean {
-    const guardIndex = cmd.indexOf(HERMES_GUARD);
-    return guardIndex >= 0 && cmd[guardIndex + 1] === action;
-  }
-
-  function isPathPreflight(cmd: string[]): boolean {
-    return (
-      cmd[0] === "python3" &&
-      cmd[1] === "-I" &&
-      cmd[2] === "-c" &&
-      typeof cmd[3] === "string" &&
-      cmd[3].includes("nemoclaw-shields-down-path-preflight")
-    );
-  }
+  const harnessFactory = createHermesUnsafeConfigHarness(requireSource, INDEX_MODULE);
+  let harness: HermesUnsafeConfigHarness;
 
   beforeEach(() => {
-    homeDir = fs.mkdtempSync(path.join(os.tmpdir(), "nemoclaw-hermes-unsafe-config-"));
-    vi.stubEnv("HOME", homeDir);
-    delete require.cache[requireSource.resolve(INDEX_MODULE)];
-    delete require.cache[requireSource.resolve("./timer-bound-lock.js")];
-    delete require.cache[requireSource.resolve("./transition-lock.js")];
-
-    const runner = requireSource("../runner.js");
-    const policy = requireSource("../policy/index.js");
-    const agentConfig = requireSource("../sandbox/agent-config.js");
-    const registry = requireSource("../state/registry.js");
-    const privilegedExec = requireSource("../sandbox/privileged-exec.js");
-    const dockerExec = requireSource("../adapters/docker/exec.js");
-    const stateDirLock = requireSource("./state-dir-lock.js");
-    const relockReconfirm = requireSource("./relock-reconfirm.js");
-    const audit = requireSource("./audit.js");
-    const permissiveRuntime = requireSource("./permissive-runtime.js");
-    const tempFiles = requireSource("../onboard/temp-files.js");
-    const childProcess = requireSource("node:child_process");
-    const timerControl = requireSource("./timer-control.js");
-    const fakeTimerPid = 4242;
-    const permissivePolicyPath = path.join(homeDir, "permissive.yaml");
-    fs.writeFileSync(permissivePolicyPath, "version: 1\nnetwork_policies: {}\n", { mode: 0o600 });
-
-    runSpy = vi.spyOn(runner, "run").mockReturnValue({ status: 0 });
-    dockerExecSpy = vi.spyOn(dockerExec, "dockerExecFileSync");
-    auditSpy = vi.spyOn(audit, "appendAuditEntry").mockImplementation(() => undefined);
-    errorSpy = vi.spyOn(console, "error").mockImplementation(() => undefined);
-    vi.spyOn(runner, "runCapture").mockReturnValue("version: 1\nnetwork_policies:\n  test: {}\n");
-    vi.spyOn(policy, "buildPolicyGetCommand").mockImplementation((name: unknown) => [
-      "policy",
-      "get",
-      String(name),
-    ]);
-    vi.spyOn(policy, "buildPolicySetCommand").mockImplementation((file: unknown, name: unknown) => [
-      "policy",
-      "set",
-      String(file),
-      String(name),
-    ]);
-    vi.spyOn(policy, "parseCurrentPolicy").mockImplementation((raw: unknown) => String(raw));
-    vi.spyOn(policy, "resolvePermissivePolicyPath").mockReturnValue(permissivePolicyPath);
-    vi.spyOn(agentConfig, "resolveAgentConfig").mockReturnValue(hermesTarget);
-    vi.spyOn(registry, "getSandbox").mockImplementation((name: unknown) => ({
-      name: String(name),
-      agent: "hermes",
-      openshellDriver: "docker",
-      lifecycleGeneration: "legacy-generation",
-      workload: { kind: "managed-image" },
-    }));
-    vi.spyOn(privilegedExec, "privilegedSandboxExecArgv").mockImplementation(
-      (_sandboxName: unknown, cmd: unknown) => cmd as string[],
-    );
-    vi.spyOn(stateDirLock, "applyStateDirLockMode").mockReturnValue([]);
-    vi.spyOn(stateDirLock, "preflightStateDirLock").mockReturnValue([]);
-    vi.spyOn(stateDirLock, "restoreStateDirLockPosture").mockReturnValue([]);
-    vi.spyOn(stateDirLock, "stateLockPlanCompatibilityIssues").mockReturnValue([]);
-    vi.spyOn(relockReconfirm, "waitForHermesInferenceRouteConvergence").mockReturnValue({
-      ok: true,
-      attempts: 1,
-      httpStatus: 200,
-    });
-    vi.spyOn(permissiveRuntime, "buildRuntimePermissivePolicy").mockImplementation(
-      (basePath: unknown) => String(basePath),
-    );
-    vi.spyOn(tempFiles, "cleanupTempDir").mockImplementation(() => undefined);
-    vi.spyOn(console, "log").mockImplementation(() => undefined);
-    vi.spyOn(console, "warn").mockImplementation(() => undefined);
-    vi.spyOn(timerControl, "readProcessStartIdentity").mockImplementation((pid: unknown) =>
-      Number(pid) === fakeTimerPid || Number(pid) === process.pid ? "test-start" : null,
-    );
-    vi.spyOn(timerControl, "isProcessAlive").mockReturnValue(true);
-    vi.spyOn(timerControl, "verifyTimerMarkerIdentity").mockReturnValue({ verified: true });
-    vi.spyOn(childProcess, "fork").mockImplementation((_module: unknown, args: unknown) => {
-      const sandboxName = String((args as unknown[])[0]);
-      return {
-        pid: fakeTimerPid,
-        disconnect: vi.fn(),
-        unref: vi.fn(),
-        kill: vi.fn(() => true),
-        send: vi.fn((message: unknown) => {
-          const request = message as { type?: unknown; processToken?: unknown };
-          if (request.type === "authorize" && typeof request.processToken === "string") {
-            const marker = timerControl.readTimerMarker(sandboxName);
-            if (marker?.timerProcessStartIdentity) {
-              fs.writeFileSync(
-                timerControl.timerAuthorizationProofPath(sandboxName, request.processToken),
-                JSON.stringify({
-                  schemaVersion: 1,
-                  pid: marker.pid,
-                  sandboxName,
-                  processToken: request.processToken,
-                  timerProcessStartIdentity: marker.timerProcessStartIdentity,
-                  authoritySha256: timerControl.timerAuthoritySha256(marker),
-                }),
-                { mode: 0o600 },
-              );
-            }
-          }
-          return true;
-        }),
-      };
-    });
-
-    dockerExecSpy.mockImplementation((cmd: string[]) => {
-      if (
-        cmd[0] === HERMES_PYTHON &&
-        cmd.includes("-c") &&
-        cmd.at(-1)?.includes("runtime-state-mutation-publisher")
-      ) {
-        return "absent";
-      }
-      if (cmd.includes(HERMES_GUARD) && cmd.includes("--help")) return CURRENT_GUARD_HELP;
-      if (isGuardAction(cmd, "begin-shields-transition")) {
-        return `lock_token=${LOCK_TOKEN} original_locked=1`;
-      }
-      if (isGuardAction(cmd, "apply-shields-transition")) {
-        return "shields_mode=mutable chattr_applied=0";
-      }
-      if (cmd[0] === "stat") {
-        return cmd.at(-1) === "/sandbox/.hermes" ? "3770 sandbox:sandbox" : "640 sandbox:sandbox";
-      }
-      if (cmd[0] === "sha256sum") return `${"b".repeat(64)}  ${cmd.at(-1)}`;
-      if (cmd[0] === "lsattr") return `---------------- ${cmd.at(-1)}`;
-      return "";
-    });
-
-    shields = requireSource(INDEX_MODULE);
+    harness = harnessFactory.beforeEachHook();
   });
 
   afterEach(() => {
-    vi.restoreAllMocks();
-    vi.unstubAllEnvs();
-    delete require.cache[requireSource.resolve(INDEX_MODULE)];
-    fs.rmSync(homeDir, { recursive: true, force: true });
+    harnessFactory.afterEachHook();
   });
 
-  function seedLockedState(sandboxName: string): string {
-    const stateDir = path.join(homeDir, ".nemoclaw", "state");
-    fs.mkdirSync(stateDir, { recursive: true });
-    fs.writeFileSync(
-      path.join(stateDir, `shields-${sandboxName}.json`),
-      JSON.stringify({
-        shieldsDown: false,
-        chattrApplied: true,
-        fileHashes: {
-          "/sandbox/.hermes/config.yaml": "b".repeat(64),
-          "/sandbox/.hermes/.env": "b".repeat(64),
-          "/sandbox/.hermes/.config-hash": "b".repeat(64),
-        },
-        updatedAt: "2026-08-11T00:00:00.000Z",
-      }),
-    );
-    return stateDir;
-  }
-
   it("rejects a Hermes config symlink before Shields down weakens posture (#8804)", () => {
-    const stateDir = seedLockedState("hermes-shields");
-    const prior = dockerExecSpy.getMockImplementation();
-    dockerExecSpy.mockImplementation((cmd: string[]) => {
-      if (isPathPreflight(cmd)) {
-        throw new Error("refusing symlink path: /sandbox/.hermes/config.yaml");
-      }
-      return prior ? prior(cmd) : "";
-    });
+    const stateDir = harness.seedLockedState("hermes-shields");
+    harness.setScenario("preflight-symlink");
 
     expect(() =>
-      shields.shieldsDown("hermes-shields", { reason: "unsafe-path", throwOnError: true }),
+      harness.shields.shieldsDown("hermes-shields", { reason: "unsafe-path", throwOnError: true }),
     ).toThrow(/refusing symlink path: \/sandbox\/\.hermes\/config\.yaml/);
 
-    expect(runSpy).not.toHaveBeenCalled();
-    expect(auditSpy).not.toHaveBeenCalled();
-    expect(
-      JSON.parse(fs.readFileSync(path.join(stateDir, "shields-hermes-shields.json"), "utf-8")),
-    ).toMatchObject({ shieldsDown: false });
-    expect(fs.existsSync(path.join(stateDir, "shields-timer-hermes-shields.json"))).toBe(false);
-    expect(shields.isShieldsDown("hermes-shields")).toBe(false);
-    expect(shields.getShieldsPosture("hermes-shields", false)).toMatchObject({
+    expect(harness.runSpy).not.toHaveBeenCalled();
+    expect(harness.auditSpy).not.toHaveBeenCalled();
+    expectHermesShieldsUpRecord(stateDir, "hermes-shields", harness.shields);
+    expect(harness.shields.getShieldsPosture("hermes-shields", false)).toMatchObject({
       locked: true,
       mutable: false,
     });
   });
 
   it("clears provisional DOWN when unlock fails on an unsafe Hermes config symlink (#8804)", () => {
-    const stateDir = seedLockedState("hermes-shields");
-    const unsafePathError = new Error("refusing to follow symlink: /sandbox/.hermes/config.yaml");
-    const prior = dockerExecSpy.getMockImplementation();
-    dockerExecSpy.mockImplementation((cmd: string[]) => {
-      if (isPathPreflight(cmd)) return "";
-      if (isGuardAction(cmd, "begin-shields-transition")) throw unsafePathError;
-      return prior ? prior(cmd) : "";
-    });
+    const stateDir = harness.seedLockedState("hermes-shields");
+    harness.setScenario("unlock-symlink");
 
     expect(() =>
-      shields.shieldsDown("hermes-shields", {
+      harness.shields.shieldsDown("hermes-shields", {
         reason: "unsafe-path",
         timeout: "15m",
         throwOnError: true,
       }),
     ).toThrow(/refusing to follow symlink: \/sandbox\/\.hermes\/config\.yaml/);
 
-    expect(
-      JSON.parse(fs.readFileSync(path.join(stateDir, "shields-hermes-shields.json"), "utf-8")),
-    ).toMatchObject({ shieldsDown: false, chattrApplied: true });
-    expect(fs.existsSync(path.join(stateDir, "shields-timer-hermes-shields.json"))).toBe(false);
-    expect(auditSpy).not.toHaveBeenCalled();
-    expect(errorSpy.mock.calls.flat().map(String).join("\n")).toContain(
+    expectHermesShieldsUpRecord(stateDir, "hermes-shields", harness.shields);
+    expect(harness.auditSpy).not.toHaveBeenCalled();
+    expect(harness.errorSpy.mock.calls.flat().map(String).join("\n")).toContain(
       "Restrictive policy restored and provisional Shields down cleared",
     );
-    expect(shields.isShieldsDown("hermes-shields")).toBe(false);
+  });
+
+  it("keeps DOWN when unlock succeeded and unsafe re-lock cannot verify protection (#8804)", () => {
+    const stateDir = harness.seedLockedState("hermes-shields");
+    harness.setScenario("unlock-ok-relock-symlink");
+    failHermesInferenceConvergence(requireSource);
+
+    expect(() =>
+      harness.shields.shieldsDown("hermes-shields", {
+        reason: "unsafe-path-after-unlock",
+        timeout: "15m",
+        throwOnError: true,
+      }),
+    ).toThrow(/Hermes inference route did not converge/);
+
+    const errors = harness.errorSpy.mock.calls.flat().map(String).join("\n");
+    expect(
+      JSON.parse(fs.readFileSync(path.join(stateDir, "shields-hermes-shields.json"), "utf-8")),
+    ).toMatchObject({ shieldsDown: true });
+    expect(errors).toContain("Manual intervention is required");
+    expect(errors).not.toContain("provisional Shields down cleared");
   });
 });
